Connections, extensions
I am rephrasing this in a far superior way; It is probably better to be providing all of this data through the pipeline. But then I think, okay, but if I ever want to change this data, then that would be a whole ordeal and a new version merged to master, or potentially a looming bug until the release is breaking. Anyway, the best design for things like this is modularity with extendability, and this new connection system implements this with a new array of extensions, which I am hoping will be able to provide a good base for these servables to work their way across. All of this being said, things are honestly getting pretty close, and I am extremely excited!