Servables 9.0: removed containers, interface redux
I completely removed containers, replaced them with the :child property. Was a lot easier, because there are no type and method errors that are going to occur with this. There are only a few key things to get done before 0.0.9, which is essentially just a pre-release for 0.1.0. Here are the major strides:
- The default files need to be upgraded, not an overly exuberant page, but a better one -- and also it should be in a module.
- The server return needs to be wrapped in some kind of way where we can continue to work with it after it is returned, giving us the live editing, 0 uptime feature -- because Toolips is just too robust. Really all we need is to get the same route dictionary that is passed through the Connection into a type and then setup the indexing.
- Documentation, including the documenter site.
- Testing, we will have to see how that will be done.