serious updates to router threading
The last thing I need to finish here is the `WorkerConnection` implementation, which will allow me to `write!` to the `Connection` from a different thread.

From there, this update is *in function*, done. I still have more work to do, as always, but within a short timeframe will come `ParametricProcesses` and `ToolipsServables` -- the precursor to this large `0.3` breaking release.