more docs :)
almost done with `core.jl` docs. After this, in no time I will be done with this project and moving onto the next step,
- `ToolipsSession`
- `ToolipsSVG`

After the completion of these modules, I have a few other `Toolips` extensions to bring up to date, but at this moment I will release `0.3.0` as the main version of `Toolips`, starting with `ToolipsServables` and finishing with `ToolipsSession` most likely. This includes ...
- `ToolipsRemote`
- `ToolipsCrawl`
- `ToolipsUploader`
- `ToolipsManager`
- `ToolipsAuth`
and `ToolipsUDP`

These will be done over a course of time -- as I continue to complete other projects in this rampage. Soon after `ToolipsSVG` as I develop these, I plan to also release the first version of `Gattino`. After all of these are finally done, I am finally going to be back onto my main project right now -- `Olive`. With this, I will also be developing `OliveCreator` and `Laboratory`. I *am* **crazy** hehe. Still far from that `Olive` moment, but `Toolips` `0.3` is a huge step towards it and we are nearly onto that step :) Lot of software is coming very soon.