updated toolipservables compat
Brought the compat binding for `ToolipsServables` up to `1.4` ... basically the change between `Toolips` `0.3.2` and `ToolipsServables` `0.1.4` is just that `gen_ref` moved to `ToolipsServables`. It seems non-release worth, I know -- but I need to get this over with so I can release `Gattino`, which I am currently writing tests for and will be releasing the first version of soon.