added `HTTPRoute`, removed default logger addition
The router generator now just checks if a logger exists, instead of adding one if it does not exist. In this case, the user will not get the link printout we have associated with `Toolips` unless a `Logger` is loaded. It still will be in the default project, so I figure most people could experiment and see the behavior I just described.